--- a/src/list_nonconcurrent.jl
+++ b/src/list_nonconcurrent.jl
@@ -20,11 +20,12 @@ function SkipList{T,M}(;
     right_sentinel = RightSentinel{T,M}(; max_height=max_height, capacity=node_capacity)
     predecessors = Vector{Node{T,M}}(undef, max_height)
     successors = Vector{Node{T,M}}(undef, max_height)
+    predecessor_offsets = Vector{Int}(undef, max_height)
     height = 1
-    length = 0
 
     for ii = 1:max_height
         link_nodes!(left_sentinel, right_sentinel, ii)
+        left_sentinel.width[ii] = 0
     end
 
     SkipList{T,M}(
@@ -33,10 +34,10 @@ function SkipList{T,M}(;
         node_capacity,
         predecessors,
         successors,
+        predecessor_offsets,
         left_sentinel,
         right_sentinel,
-        height,
-        length
+        height
     )
 end
 
@@ -46,12 +47,20 @@ SkipList external API
 
 height(list::SkipList) = list.height
 
-Base.length(list::SkipList) = list.length
+Base.length(list::SkipList) = list.left_sentinel.width[list.height]
 
 node_capacity(list::SkipList) = list.node_capacity
 
+Base.size(list::SkipList) = (length(list),)
+Base.IndexStyle(::Type{SkipList}) = IndexLinear()
+
+function Base.getindex(list::SkipList, i)
+    node, k = find_index(list, i)
+    node.vals[k]
+end
+
 function Base.insert!(list::SkipList{T,M}, val) where {T,M}
-    level_found, predecessors, successors = find_node(list, val)
+    level_found, predecessors, successors, predecessor_offsets = find_node(list, val)
 
     # We insert into the predecessor node from our search that sits in the
     # first level of the skip list. If that node is full, we must split it
@@ -74,19 +83,32 @@ function Base.insert!(list::SkipList{T,M}, val) where {T,M}
                     p=height_p(list),
                 )
 
-                # If the new node has height greater than the old height of the list,
-                # we must increase the size of the list.
-                if height(list) < height(new_node)
-                    for ii = height(list)+1:height(new_node)
+                # Fix node widths to account for elements removed from `old_node`
+                # (`new_node` is still not in the list)
+                n_removed = length(new_node)
+                for level = 1:height(list)
+                    @assert level > height(old_node) || predecessors[level] === old_node
+                    predecessors[level].width[level] -= n_removed
+                end
+
+                # Height of list should be kept at least one higher than highest non-sentinel node
+                # (since the width of the left sentinel must be correctly maintained at the 
+                #  level above the highest non-sentinel).
+                # Grow as necessary before inserting the new node.
+                if height(list) < height(new_node) + 1
+                    len = length(list) # == list.left_sentinel.width[height(list)]
+                    for ii = height(list)+1:height(new_node)+1
                         predecessors[ii] = list.left_sentinel
                         successors[ii] = list.right_sentinel
+                        predecessor_offsets[ii] = 0
+                        list.left_sentinel.width[ii] = len
                     end
 
-                    list.height = height(new_node)
+                    list.height = height(new_node) + 1
                 end
 
                 # Insert the new node between the predecessor / successor nodes.
-                interpolate_node!(predecessors, successors, new_node)
+                interpolate_node!(list, new_node, predecessors, successors, predecessor_offsets)
 
                 # We insert into the right node if its key is ≤ the insertion value;
                 # otherwise, we insert into the left node.
@@ -97,7 +119,12 @@ function Base.insert!(list::SkipList{T,M}, val) where {T,M}
         end
 
         insert!(insertion_node, val)
-        list.length += 1
+        for level = 1:height(insertion_node)
+            insertion_node.width[level] += 1
+        end
+        for level = height(insertion_node)+1:height(list)
+            predecessors[level].width[level] += 1
+        end
 
         Some(val)
     end
@@ -116,15 +143,20 @@ function Base.delete!(list::SkipList, val)
 
     deletion_node = successors[1]
     result = delete!(deletion_node, val)
+    for ii = 1:height(deletion_node)
+        deletion_node.width[ii] -= 1
+    end
+    for ii = height(deletion_node)+1:height(list)
+        predecessors[ii].width[ii] -= 1
+    end
 
     # If the node is empty, we must delete it and remove it from the list
     if isempty(deletion_node) && !is_sentinel(deletion_node)
         for ii = 1:height(deletion_node)
             link_nodes!(predecessors[ii], next(deletion_node, ii), ii)
+            predecessors[ii].width[ii] += deletion_node.width[ii]
         end
     end
-
-    list.length -= 1
 
     Some(val)
 end
@@ -165,17 +197,20 @@ function find_node(list::SkipList{T,M}, val; right_if_member = true) where {T,M}
     list_height = height(list)
     predecessors = list.predecessor_buffer
     successors = list.successor_buffer
+    predecessor_offsets = list.predecessor_offset_buffer
 
     level_found = NODE_NOT_FOUND
 
-    current_node = list.left_sentinel
+    current_node, current_offset = list.left_sentinel, 0
     for ii = list_height:-1:1
         # Move to the right until we reach a node whose key is
         # greater than the value we're searching for
         next_node = next(current_node, ii)
+        next_offset = current_offset + current_node.width[ii]
         while right_if_member ? (next_node < val || val ∈ next_node) : (next_node < val && val ∉ next_node)
-            current_node = next_node
+            current_node, current_offset = next_node, next_offset
             next_node = next(current_node, ii)
+            next_offset = current_offset + current_node.width[ii]
         end
 
         # Note to future self: trying to figure out how to delete node. Need
@@ -190,8 +225,43 @@ function find_node(list::SkipList{T,M}, val; right_if_member = true) where {T,M}
         end
         predecessors[ii] = current_node
         successors[ii] = next_node
+        predecessor_offsets[ii] = current_offset
     end
 
-    level_found, predecessors, successors
+    level_found, predecessors, successors, predecessor_offsets
 end
+
+"""
+Find `node` and `k` such that `node.vals[k]` is the `i`th list element
+"""
+function find_index(list::SkipList, i)
+    @boundscheck Base.checkbounds(list, i)
+    offset = 0
+    node = list.left_sentinel
+    for level = height(list):-1:1
+        while (next_offset = offset + node.width[level]) < i
+            node = next(node, level)
+            offset = next_offset
+        end
+    end
+    node, i-offset
+end
+
+"""
+Insert a new node between a list of predecessor and successor nodes.
+"""
+function interpolate_node!(list, node, predecessors, successors, predecessor_offsets)
+    offset = predecessor_offsets[1] + predecessors[1].width[1]
+    for level = 1:height(node)
+        link_nodes!(predecessors[level], node, level)
+        link_nodes!(node, successors[level], level)
+        d = offset - predecessor_offsets[level]
+        node.width[level] = predecessors[level].width[level] - d + length(node)
+        predecessors[level].width[level] = d
+    end
+    for level = height(node)+1:height(list)
+        predecessors[level].width[level] += length(node)
+    end
+end
+