--- a/src/node_nonconcurrent.jl
+++ b/src/node_nonconcurrent.jl
@@ -18,7 +18,8 @@ function Node{T,M}(
 
     height = min(height, max_height)
     next = Vector{Node{T,M}}(undef, height)
-    Node{T,M}(vals, next, capacity, flags)
+    width = fill(length(vals), height)
+    Node{T,M}(vals, next, width, capacity, flags)
 end
 
 LeftSentinel{T,M}(; max_height = DEFAULT_MAX_HEIGHT, kws...) where {T,M} =
@@ -44,9 +45,7 @@ capacity(node::Node) = node.capacity
 Base.isempty(node::Node) = (length(node) == 0)
 isfull(node::Node) = is_sentinel(node) || (length(node) == capacity(node))
 
-Base.in(val, node::Node) =
-    searchsorted(node.vals, val) |>
-    idx -> first(idx) ≤ last(idx)
+Base.in(val, node::Node) = insorted(val, node.vals)
 
 @generated function Base.insert!(node::Node{T,M}, val) where {T,M}
     quote
@@ -84,16 +83,6 @@ function split!(node::Node{T,M}; kws...) where {T,M}
     median_idx = div(length(node), 2)
     right_vals = splice!(node.vals, median_idx+1:length(node))
     node, Node{T,M}(right_vals; kws...)
-end
-
-"""
-Insert a new node between a list of predecessor and successor nodes
-"""
-function interpolate_node!(predecessors, successors, node)
-    for level = 1:height(node)
-        link_nodes!(predecessors[level], node, level)
-        link_nodes!(node, successors[level], level)
-    end
 end