--- a/src/node_nonconcurrent.jl
+++ b/src/node_nonconcurrent.jl
@@ -33,8 +33,14 @@ Node external API
 ===========================#
 
 function Base.string(node::Node)
-    result = "$(node.vals), height = $(height(node))"
-    "Node($result)"
+    if is_left_sentinel(node)
+        "<left sentinel>"
+    elseif is_right_sentinel(node)
+        "<right sentinel>"
+    else
+        result = "$(node.vals), height = $(height(node))"
+        "Node($result)"
+    end
 end
 
 Base.length(node::Node) = length(node.vals)
@@ -46,38 +52,6 @@ Base.isempty(node::Node) = (length(node) == 0)
 isfull(node::Node) = is_sentinel(node) || (length(node) == capacity(node))
 
 Base.in(val, node::Node) = insorted(val, node.vals)
-
-@generated function Base.insert!(node::Node{T,M}, val) where {T,M}
-    quote
-        if length(node) ≥ capacity(node)
-            "Node size exceeds capacity ($(capacity(node)))" |>
-            ErrorException |>
-            throw
-        end
-
-        idx = searchsorted(node.vals, val)
-
-        $(
-            if M == :Set
-                quote
-                    if first(idx) > last(idx)     # Value not found in node
-                        insert!(node.vals, first(idx), val)
-                    end
-                end
-            else
-                quote
-                    insert!(node.vals, first(idx), val)
-                end
-            end
-        )
-    end
-end
-
-Base.delete!(node::Node, val) =
-    searchsorted(node.vals, val) |>
-    idx -> if first(idx) ≤ last(idx)
-        deleteat!(node.vals, first(idx))
-    end
 
 function split!(node::Node{T,M}; kws...) where {T,M}
     median_idx = div(length(node), 2)