This PR:
 - adds O(log(n)) indexing support to `SkipLists` (i.e., accessing the `i`th smallest element) (I'm using this to implement fast running median with correct NaN handling).
 - optimizes by avoiding the double call to `searchsorted` in insertion and deletion (once to find the correct node, and again to insert/delete in the node).

Result of the `benchmark_generate_skiplist` benchmark from `benchmark/benchmark_generation.jl`:
- version 1.1.0:
```
 BenchmarkTools.Trial: 73 samples with 1 evaluation.
 Range (min … max):  50.266 ms … 83.954 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     58.545 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   58.636 ms ±  4.699 ms  ┊ GC (mean ± σ):  1.03% ± 2.36%

             ▂    ▂  ▅   ▂ ▅  ▅  █ ▂     ▂▂
  ▅▁▁▅▁▁▁█▅▅▁█▅██▅██▅█▁▅▁███▁▁█▅▅█▅██▁█▅▅████▁█▁█▅▁▁▁▁▁▁▁▁▁▁▅ ▁
  50.3 ms         Histogram: frequency by time        67.7 ms <

 Memory estimate: 6.34 MiB, allocs estimate: 1597.
```
- this PR:
```
BenchmarkTools.Trial: 109 samples with 1 evaluation.
 Range (min … max):  31.594 ms … 51.328 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     35.377 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   36.026 ms ±  2.934 ms  ┊ GC (mean ± σ):  1.83% ± 3.98%

       ▃  █ ▅▃  ▂▂▂
  ▃▁▁▅▆█▆▇█▇██▆▇███▁▃▇█▃▁▇▇▁▃▃▇▃▃▃▅▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃ ▃
  31.6 ms         Histogram: frequency by time        48.5 ms <

 Memory estimate: 6.35 MiB, allocs estimate: 2107.
```