@kernelmethod 2022-06-10
Thanks for the PR! I'll take a look at this over the weekend, but from an initial skim over this looks great.