@kernelmethod 2022-06-26
On the whole this looks pretty good. In the long term I might need to make some small changes -- I'd forgotten what a mess this library was -- but for now I'll merge this in alongside a few other changes and create a new release.

Thanks again!