@yha 2022-06-28
Thanks!
New version of `RunningQuantiles.jl` speeded up by indexable SkipLists: https://github.com/JuliaRegistries/General/pull/63289